Nuevo: Preferencia, Modificados: Comuna,Match,Region,Usuario,Perfil

+++ src/main/java/com/loveu/loveu/model/Match.java
@@ -27,13 +27,13 @@ import lombok.NoArgsConstructor;
 public class Match {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "user1_id", nullable = false)
-    private Long user1Id;
+    private Integer user1Id;
 
     @Column(name = "user2_id", nullable = false)
-    private Long user2Id;
+    private Integer user2Id;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
